Set the `env` property to align with mixin
These deploy configs set a value in the env block in the deploy config
that is different from the value that is set in the mixin. Because the
mixin overrides the value that is set in the deploy config, these values
are not currently respected in the effective deploy config.

However, Build and Deploy is going to change the
mixin-overrides-deploy-config behavior to the more intuitive behavior
where the deploy config is preferred over the mixin where they conflict.
This PR changes your deploy config so that after the change to the
mixins takes place, the effective deploy config will remain the same.

If you have any questions, please leave a comment on HubSpot/Orion#745
or ask Build and Deploy in #platform-support

/cc @HubSpot/paas-build-and-deploy