Update `postDeployTests` format
This was in the wrong format before; it needs to be [environment]>mesos,
rather than vice versa.

This wasn't correct before either (you can check that it wasn't
launching acceptance tests before eg
https://private.hubteam.com/orion/deploy-history/8148067) but we never
noticed because Orion hadn't been configured to check the presence of
unknown fields.

That was changed a couple of months ago, but would only take place after
a deploy config was updated. This was updated recently-ish to change the
team, which made Orion re-read the deploy config and notice that this
was not valid.