Pick up BuildScripts change to eagerly create cache
Pick up a change in BuildScripts to eagerly create the directory to be
cached. At the moment, if there is no existing cache and nothing in the
build writes to the cache, the cache-target directory is never created,
which means that packaging the cache fails, which means the build fails.
Since that's bad - although hopefully a degenerate case - let's try to
avoid that.