Reduce frequency of QA db cleaner
This cleans the QA database, which is only used for Orion-branch QA
deploys. This job also really only handles repositories that are
archived; deploy configs that are deleted in other ways (should be)
handled by the webhooks from GitHub.

Given that, we can probably drastically reduce the frequency at which
this job runs, from hourly to daily. This job still accounts for 10k
req/hr, so this should change it from 10k req/hr to 10k req/day.

5am UTC converts to 1am current Cambridge time, 6a current Dublin time
(with some variance for daylight savings time), so this should be
outside of work hours for Cambridge & Dublin.